libmbim: Add mbim-proxy support
This patch implements the mbim-proxy support to allow mulitple clients
to have concurrent access to the MBIM device through libmbim.

The implementation is heavily based on qmi-proxy.

The function mbim_device_open_full() is provided for clients to specify that
the device should be opened through the mbim-proxy via the flag
MBIM_DEVICE_OPEN_FLAGS_PROXY.

The function mbim_device_open() will behave as it has before and open the
device directly.

Proxy Control:
A new service UUID is created to allow for clients to control the proxy.
This service is currently only used to pass on the device path (/dev/cdc-wdm*)
the client is requesting to open. In the future CIDs will be added to allow
for setting the MaxControlMessage size and setting timeout for specific
messages.

Unknown UUIDs:
Currently the mbim-proxy will allow messages with unknown UUIDs to pass through.
The proxy will track unknown UUIDs sent by the client and device notifications
matching the tracked UUID will be sent to the client.  Device notifications
of known UUIDs will be sent to all connected clients.

Signed-off-by: Greg Suarez <gsuarez@smithmicro.com>