libmbim-glib,proxy: Allow for --no-close and --no-open options by mbimcli
To support the --no-close option the proxy will no longer send the close
message to the device. The proxy will also only exit once all devices
are removed.  Thus the proxy connects to the MBIM_DEVICE_SIGNAL_REMOVED
signal for each device. A new property, MBIM_PROXY_N_DEVICES, was added
to indicate the number of devices the proxy is managing.

To support the --no-open option the proxy now calls the
mbim_device_new() upon receiving the
MBIM_CID_PROXY_CONTROL_CONFIGURATION with the device path.